[const] Centralize constants for overrides
